only define the mongo connection string once, in a utility function
