+++ src/index.js
@@ -1,14 +1,17 @@
-const express = require('express')
-const config = require('config')
 const cors = require('cors')
-const morgan = require('morgan')
-const mongoose = require('mongoose')
-const helmet = require('helmet')
-const mw = require('./middleware/middleware')
-const logger = require('./middleware/logger')
-const configureRoutes = require('./routes.config')
+const config = require('config')
+const express = require('express')
 const app = express()
+const helmet = require('helmet')
+const mongoose = require('mongoose')
+const morgan = require('morgan')
+
+const configureRoutes = require('./routes.config')
+const dbUtils = require('./utils/db')
 const errors = require('./utils/error')
+const logger = require('./middleware/logger')
+const mw = require('./middleware/middleware')
+
 const error = new errors.IDRError()
 require('dotenv').config() // This enables us to read from the .env file.
 
@@ -40,30 +43,8 @@ app.use((req, res, next) => {
   res.status(404).json(error.notFound())
 })
 
-// construct MongoDB connection string
-// assumes that host, port, database are always defined in default config, but
-// that username and password may not be
-const appEnv = process.env.NODE_ENV
-let dbUser, dbPassword
-
-if (process.env.MONGO_USER && process.env.MONGO_PASSWORD) {
-  dbUser = process.env.MONGO_USER
-  dbPassword = process.env.MONGO_PASSWORD
-} else {
-  dbUser = config.has(`${appEnv}.username`) ? config.get(`${appEnv}.username`) : false
-  dbPassword = config.has(`${appEnv}.password`) ? config.get(`${appEnv}.password`) : false
-}
-
-const dbHost = process.env.MONGO_HOST ? process.env.MONGO_HOST : config.get(`${appEnv}.host`)
-const dbPort = process.env.MONGO_PORT ? process.env.MONGO_PORT : config.get(`${appEnv}.port`)
-const dbName = config.get(`${appEnv}.database`)
-const dbLoginPrepend = (dbUser && dbPassword) ? `${dbUser}:${dbPassword}@` : ''
-const dbConnectionStr = `mongodb://${dbLoginPrepend}${dbHost}:${dbPort}/${dbName}`
-
-logger.info(`Using NODE_ENV '${process.env.NODE_ENV}' and app environment '${appEnv}'`)
-logger.info('Using dbName = ' + config.get(`${appEnv}.database`))
-
 // Connect to MongoDB database
+const dbConnectionStr = dbUtils.getMongoConnectionString()
 mongoose.connect(dbConnectionStr, {
   useNewUrlParser: true,
   useUnifiedTopology: true,
