+++ src/scripts/hashPasswords.js
@@ -4,17 +4,17 @@
  * in the database
  */
 
-const express = require('express')
-const config = require('config')
-const mongoose = require('mongoose')
-const app = express()
-const logger = require('../middleware/logger')
-const User = require('../model/user')
 const argon2 = require('argon2')
+const express = require('express')
+const app = express()
+const mongoose = require('mongoose')
 const uuidAPIKey = require('uuid-apikey')
+
+const logger = require('../middleware/logger')
 const errors = require('../utils/error')
+const User = require('../model/user')
+
 const error = new errors.IDRError()
-require('dotenv').config() // This enables us to read from the .env file.
 
 // Body Parser Middleware
 app.use(express.json()) // Allows us to handle raw JSON data
@@ -22,27 +22,8 @@ app.use(express.urlencoded({ extended: false })) // Allows us to handle url enco
 // Make mongoose connection available globally
 global.mongoose = mongoose
 
-// construct MongoDB connection string
-// assumes that host, port, database are always defined in default config, but
-// that username and password may not be
-const appEnv = process.env.NODE_ENV
-let dbUser, dbPassword
-
-if (process.env.MONGO_USER && process.env.MONGO_PASSWORD) {
-  dbUser = process.env.MONGO_USER
-  dbPassword = process.env.MONGO_PASSWORD
-} else {
-  dbUser = config.has(`${appEnv}.username`) ? config.get(`${appEnv}.username`) : false
-  dbPassword = config.has(`${appEnv}.password`) ? config.get(`${appEnv}.password`) : false
-}
-
-const dbHost = process.env.MONGO_HOST ? process.env.MONGO_HOST : config.get(`${appEnv}.host`)
-const dbPort = process.env.MONGO_PORT ? process.env.MONGO_PORT : config.get(`${appEnv}.port`)
-const dbName = config.get(`${appEnv}.database`)
-const dbLoginPrepend = (dbUser && dbPassword) ? `${dbUser}:${dbPassword}@` : ''
-const dbConnectionStr = `mongodb://${dbLoginPrepend}${dbHost}:${dbPort}/${dbName}`
-
 // Connect to MongoDB database
+const dbConnectionStr = dbUtils.getMongoConnectionString()
 mongoose.connect(dbConnectionStr, {
   useNewUrlParser: true,
   useUnifiedTopology: false,
