+++ src/scripts/populate.js
@@ -3,13 +3,12 @@
  * database with static fixtures at `cve-services/datadump/pre-population`
  */
 
-const config = require('config')
-
 const express = require('express')
 const app = express()
 const mongoose = require('mongoose')
 
 const dataUtils = require('../utils/data')
+const dbUtils = require('../utils/db')
 const errors = require('../utils/error')
 const logger = require('../middleware/logger')
 const CveIdRange = require('../model/cve-id-range')
@@ -19,8 +18,6 @@ const Org = require('../model/org')
 const User = require('../model/user')
 
 const error = new errors.IDRError()
-
-require('dotenv').config() // This enables us to read from the .env file.
 
 const populateTheseCollections = {
   'Cve': Cve,
@@ -36,27 +33,8 @@ app.use(express.urlencoded({ extended: false })) // Allows us to handle url enco
 // Make mongoose connection available globally
 global.mongoose = mongoose
 
-// construct MongoDB connection string
-// assumes that host, port, database are always defined in default config, but
-// that username and password may not be
-const appEnv = process.env.NODE_ENV
-let dbUser, dbPassword
-
-if (process.env.MONGO_USER && process.env.MONGO_PASSWORD) {
-  dbUser = process.env.MONGO_USER
-  dbPassword = process.env.MONGO_PASSWORD
-} else {
-  dbUser = config.has(`${appEnv}.username`) ? config.get(`${appEnv}.username`) : false
-  dbPassword = config.has(`${appEnv}.password`) ? config.get(`${appEnv}.password`) : false
-}
-
-const dbHost = process.env.MONGO_HOST ? process.env.MONGO_HOST : config.get(`${appEnv}.host`)
-const dbPort = process.env.MONGO_PORT ? process.env.MONGO_PORT : config.get(`${appEnv}.port`)
-const dbName = config.get(`${appEnv}.database`)
-const dbLoginPrepend = (dbUser && dbPassword) ? `${dbUser}:${dbPassword}@` : ''
-const dbConnectionStr = `mongodb://${dbLoginPrepend}${dbHost}:${dbPort}/${dbName}`
-
 // Connect to MongoDB database
+const dbConnectionStr = dbUtils.getMongoConnectionString()
 mongoose.connect(dbConnectionStr, {
   useNewUrlParser: true,
   useUnifiedTopology: false,
@@ -71,7 +49,7 @@ db.on('error', () => {
 })
 
 db.once('open', async () => {
-  logger.info(`Successfully connected to database ${dbName} at ${dbHost}:${dbPort}`)
+  logger.info('Successfully connected to database!')
 
   // script runner (currently) needs to agree to an action that drops collections
   const userInput = dataUtils.getUserPopulateInput()
