+++ src/utils/data.js
@@ -91,15 +91,15 @@ async function newCveIdTransform (cveId) {
   return cveId
 }
 
-function getUserPopulateInput (appName, dbName) {
+function getUserPopulateInput (appName, dbName, collectionNames) {
   let userInput = prompt(
-    `Are you sure you wish to pre-populate the database for the ${appName} environment?` +
+    `Are you sure you wish to pre-populate the database for the ${appName} environment? ` +
     `Doing so will drop the 'Cve' collection in the ${dbName} database. (y/n) `
   )
   while (userInput.toLowerCase() !== 'n' && userInput.toLowerCase() !== 'y') {
     console.log('Unrecognized Input')
     userInput = prompt(
-      `Do you wish to pre-populate the database for the ${appName} environment?` +
+      `Do you wish to pre-populate the database for the ${appName} environment? ` +
       `Doing so will drop the 'Cve', collection in the ${dbName} database. (y/n) `
     )
   }
